refactor(lib): 产物激活按规划/时新性/输入认领/注册四职责拆分
